fix(amazonq): auto-review removes existing issues (#6535)
## Problem

Auto-reviews often produce less code issues than manual reviews, but the
current behavior is to remove all the issues in the file when processing
the new ones. This means that issues discovered by manual reviews but
not auto-reviews will silently disappear if auto-reviews is enabled.


## Solution

- Auto-reviews should not clear the previous issues, but instead merge
in the new results to the existing group.
- Fixed a related issue with the `ignoreIssue` command being flaky


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.