telemetry(auth): Update metrics to better debug Auth dropoff (#6625)
## Problem

We noticed that there was an auth dropoff between `session_start` with a
brand new clientId, versus when the `auth_userState` metric indicated
`isFirstUse` meaning the user is net new. We went from 12.7k for
`session_start` to 9.8k for `auth_userState`, these should have been
basically the same.

## Solution

Add in certain metrics to help debug where the discrepancy is coming
from:

- When we determine the user is a first time user, we will also check if
the clientId is newly generated. If this is not the case we know there
is a discrepancy here
- The relevant metric will be `function_call` with a `functionName:
isFirstUse`, `result: Failed`, and a `reason: ClientIdAlreadyExisted`
- When we determine the user is a first time user, if we detected that
they had previous auth connections, this will indicate a likely cause
for the discrepancy
- The relevant metric will be `function_call` with `reason:
UnexpectedConnections`
- We will emit metrics when the Auth Login page loads since that also
had a discrepancy and the telemetry did not exist
- The relevant metric is `webview_load` and it will indicate when the
Auth Login/Reauth page has actually loaded
- Previously we were observing the telemetry for the command
`aws.amazonq.focusChat`, but all this did was emit when called and
didn't confirm the UI actually loaded.
- We will also add the `isFirstUse` metric source value in to some other
existing metrics

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>