merge (#28)
* chore: merge release candidate version rc-20250815 (#7916)

## Problem
This merges the released changes for rc-20250815 into main.
MCM-132479224

## Solution


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Lei Gao <97199248+leigaol@users.noreply.github.com>
Co-authored-by: Na Yue <yueny@amazon.com>
Co-authored-by: Will Lo <96078566+Will-ShaoHua@users.noreply.github.com>
Co-authored-by: Aidan Ton <atona@amazon.com>
Co-authored-by: atontb <104926752+atonaamz@users.noreply.github.com>

* fix(sagemaker): Reduce number of reconnect retries (#7827)

## Problem
When a connection to a remote space is severed, the extension will
attempt to reconnect up to 60 times. Each reconnect attempt opens the
Space Details page in a new browser tab. Opening up to 60 new tabs is a
poor user experience.

## Solution
We change the maximum number of retries to 8.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Co-authored-by: Newton Der <dernewtz@amazon.com>

* fix(amazonq): do not auto trigger if the ongoing typeahead does not match (#7924)

## Problem

Bug: When inline suggestion is active but user types a character that
does not match the typeahead, this typing event triggered a inline
completion API call that eventually did not manage to display suggestion
on screen, which eventually reported as rejected completions.






## Solution


Previously, before Flare migration, we had this design:


https://github.com/aws/aws-toolkit-vscode/blob/amazonq/v1.74.0/packages/amazonq/src/app/inline/activation.ts#L108

Auto trigger will be disabled if at that moment the suggestion is
visible.

Under the previous design, the first character that does not match the
typeahead will not make a new auto trigger, it simply rejects the
current inline completion.

This PR is to bring back the old design.

It is not customer facing, hence no change log.


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: tsmithsz <84354541+tsmithsz@users.noreply.github.com>
Co-authored-by: Na Yue <yueny@amazon.com>
Co-authored-by: Will Lo <96078566+Will-ShaoHua@users.noreply.github.com>
Co-authored-by: Aidan Ton <atona@amazon.com>
Co-authored-by: atontb <104926752+atonaamz@users.noreply.github.com>
Co-authored-by: Newton Der <newton.der@gmail.com>
Co-authored-by: Newton Der <dernewtz@amazon.com>